Move Avatar_Privacy\Data_Storage\Database to A..\Database\Table
