+++ includes/avatar-privacy/components/class-setup.php
@@ -34,7 +34,7 @@ use Avatar_Privacy\Core\User_Fields;
 
 use Avatar_Privacy\Components\Image_Proxy;
 
-use Avatar_Privacy\Data_Storage\Database;
+use Avatar_Privacy\Data_Storage\Database\Table as Database;
 use Avatar_Privacy\Data_Storage\Network_Options;
 use Avatar_Privacy\Data_Storage\Options;
 use Avatar_Privacy\Data_Storage\Site_Transients;
